+++ erudit/models/core.py
@@ -594,6 +594,34 @@ class Article(EruditDocument, FedoraMixin, FedoraDated):
         verbose_name_plural = _("Articles")
 
 
+class Thesis(EruditDocument):
+    """ Represents a single thesis. """
+    collection = models.ForeignKey(Collection, verbose_name=_('Collection'))
+    """ The collection associated with the considered thesis. """
+
+    author = models.ForeignKey('Author', verbose_name=_('Auteur'))
+    """ The author associated with the considered thesis. """
+
+    title = models.CharField(max_length=600, verbose_name=_('Titre'))
+    """ The title of the thesis. """
+
+    url = models.URLField(verbose_name=_('URL'))
+    """ The URL of the considered thesis. """
+
+    publication_year = models.PositiveIntegerField(verbose_name=_('Année de publication'))
+    """ The publication year of the thesis. """
+
+    description = models.TextField(verbose_name=_('Résumé'), blank=True, null=True)
+    """ A thesis can have a description. """
+
+    class Meta:
+        verbose_name = _('Thèse')
+        verbose_name_plural = _('Thèses')
+
+    def __str__(self):
+        return self.title
+
+
 class Author(Person):
     suffix = models.CharField(max_length=50, verbose_name=_('Suffixe'), blank=True, null=True)